pyproject.toml
[project]
name = "orange-hrm-playwright"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "allure-pytest>=2.15.3",
    "playwright>=1.58.0",
    "pytest-html>=4.2.0",
    "pytest-playwright>=0.7.2",
]
[tool.pylint.messages_control]
disable = ["missing-module-docstring", "missing-function-docstring"]
